Moved tests to src/tests
