+++ src/tests/App.test.js
@@ -1,8 +1,8 @@
 import React from "react";
 import ReactDOM from "react-dom";
 import { Provider } from "react-redux";
-import App from "./App";
-import store from "./store";
+import App from "../App";
+import store from "../store";
 
 beforeEach(() => {
   window.getSelection = () => {
