+++ src/tests/Toolbar.test.js
@@ -1,7 +1,7 @@
 import React from "react";
 import Enzyme, { shallow } from "enzyme";
 import Adapter from "enzyme-adapter-react-16";
-import Toolbar from "./Toolbar";
+import Toolbar from "../components/editor/toolbar/Toolbar";
 
 Enzyme.configure({ adapter: new Adapter() });
